.cc file to perform operations and a few other useful utilities for main_with_Hash_Table.cc
This file performs the main functionality of performing the operations once the string has been parsed and all chars(operations) and ints(numbers) have been put into the deques. This .cc will eventually be edited to get rid of the main() function and then compiled separately with the actual main so that the process of performing equations may be abstracted away.