I rewrote main but it's not complete by far
All it does at this point is accept and create key/value pairs in the hash table. When input is like : LET X = 11 : format i mean. 

The inspiration to parse it this way came from the Bitfield project, this project seems similar in many ways to Bitfield and RPN Calculator imo. 

My next goal is to create a method to accept the variables, numbers, and operators and put them into the int(number) deque and char(operators) deque. Once that is done it should be straight-forward to implement with the separated equation functionality, but don't quote me on that :)