Final code.
This code passes 16(80%) of the test cases, throws exceptions, and has O(1) lookup/insert/delete. Professor Kerney approved the LOOP functionality and said the commenting was sufficient. I might go back and make it pass more tests for fun if I have time but it IS done at this point.